kbs: Add nebula_ca plugin
This plugin calls the nebula-cert binary to provide some of its
functionality for Nebula nodes (e.g. CoCo PODs or confidential VMs) that
want to join an encrypted Nebula overlay network (ENON).

The Nebula CA is created when the KBS is started.

Currently, the only runtime service it provides is the credential
service which returns the following structure for the caller to join an
ENON.

struct CredentialServiceOut {
    node_crt: Vec<u8>,  // Self-signed certificate created
    node_key: Vec<u8>,  // Key created
    ca_crt: Vec<u8>,    // CA certificate
}

Since the CredentialServiceOut is returned only after the caller is
attested, the ENON is expected to have only attested nodes.

Some use cases for an ENON:
- The workload does not trust the network interface provided by the CSP
- The workload does not support data encryption over the network

For further information on configuration or usage, please refer to the
kbs/docs/plugins/nebula_ca.md or the official Nebula website
https://github.com/slackhq/nebula

Signed-off-by: Claudio Carvalho <cclaudio@linux.ibm.com>